Add some Gestalt manager
